Parse the registration response in u2ftoken
Add a new u2ftoken.Token.RegisterParsed method which takes the raw
result of u2ftoken.Token.Register and parses it according to spec.

The format is described at:
https://fidoalliance.org/specs/fido-u2f-v1.2-ps-20170411/fido-u2f-raw-message-formats-v1.2-ps-20170411.html#registration-response-message-success

Parsing the registration response is necessary to extract the key
handle for use in future authentication. It's also recommended to verify
the response signature and attestation certificate.